requirements.txt

Django==2.0.2
djangorestframework==3.7.7
django_redis==4.9.0
django-constance==2.1.0
django-extensions==2.0.2
django-cors-headers==2.2.0

requests==2.18.4
websocket-client==0.46.0
ecdsa==0.13
pycrypto==2.6.1
base58==0.2.5
msgpack==0.5.6

-e git+https://github.com/aeternity/aepp-sdk-python.git#egg=aeternity